requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.3
pymongo
redis
python-dotenv
yt-dlp
git+https://github.com/pytgcalls/pytgcalls.git
Pillow
